added usize size check and simplified memcase

+++ epserde/src/lib.rs
@@ -28,7 +28,7 @@ pub use des::*;
 pub use ser::*;
 
 /// (Major, Minor) version of the file format, this follows semantic versioning
-pub const VERSION: (u32, u32) = (0, 0);
+pub const VERSION: (u16, u16) = (0, 0);
 
 /// Magic + endianess marker
 pub const MAGIC: u64 = u64::from_ne_bytes(*b"epserde ");
